// src/lib/rules/image-optimization-cloudinary/cloudinary-types.ts
/* eslint-disable camelcase */
import { IAsyncHTMLElement } from '../../types';

export type cloudinaryResult = {
    bytes: number;
    element: IAsyncHTMLElement;
    format: string;
    height: number;
    originalBytes: number;
    originalUrl: string;
    public_id: string;
    url: string;
    version: number;
    width: number;
};
/* eslint-enable camelcase */